move StepsController inside the modal body
