+++ src/containers/permissions/assign-permissions-to-data-set/index.js
@@ -2,4 +2,4 @@
  * @flow
  */
 
-export { default as AssignPermissionsToDataSet } from './AssignPermissionsToDataSet';
+export { default as AssignPermissionsToDataSetModalBody } from './AssignPermissionsToDataSetModalBody';
